--- src/steamship/experimental/transports/__init__.py
@@ -1,4 +1,0 @@
-from .telegram import TelegramTransport
-from .transport import Transport
-
-__all__ = ["Transport", "TelegramTransport"]
